alternatives/server-flask/test/conftest.py
import pytest
from blog import create_app
from blog.config import TestConfig


@pytest.fixture
def app():
    app = create_app(TestConfig)
    yield app


@pytest.fixture
def client(app):
    return app.test_client()
