alternatives/server-flask/requirements-prod.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-prod.in
#
blinker==1.4              # via sentry-sdk
boto3==1.9.220
botocore==1.12.220        # via boto3, s3transfer
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.0                # via flask
docutils==0.15.2          # via botocore
flask==1.1.1
google-api-core[grpc]==1.14.2  # via google-cloud-core, google-cloud-logging
google-auth==1.6.3        # via google-api-core
google-cloud-core==1.0.3  # via google-cloud-logging
google-cloud-logging==1.12.1
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.23.0            # via google-api-core
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
protobuf==3.9.1           # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.3
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
pydantic==0.32.2
python-dateutil==2.8.0    # via botocore
pytz==2019.2              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
s3transfer==0.2.1         # via boto3
sentry-sdk[flask]==0.11.2
six==1.12.0               # via google-api-core, google-auth, grpcio, protobuf, python-dateutil
urllib3==1.25.3           # via botocore, requests, sentry-sdk
uwsgi==2.0.18
uwsgidecorators==1.1.0
werkzeug==0.15.5          # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via google-api-core, protobuf
